Unconditionally evaluate `hook.extra` as a function
In nixos/nixpkgs#386208, function arguments are preserved in some
circumstances. This caused issues such as divnix/std#399. Ultimately,
it comes down to the `builtins.isFunction` check, which does not detect
functors.

With this patch, `hook.extra` is coerced to a function unconditionally.
If it is a constant, any arguments are ignored, so behaviour does not
change.